requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langgraph>=0.1.0
langchain-groq>=0.1.9
langchain-core>=0.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
